feat : 프로퍼티 추론이 가능하도록 productElement 인터페이스 추가
