+++ src/util/functions/create-product-pagination-form.function.ts
@@ -1,7 +1,7 @@
 import { GetProductListPaginationDto } from 'src/entities/dtos/get-product-list-pagination.dto';
 import { GetProductListResponse } from 'src/interfaces/get-product-list-response.interface';
 import { GetResponse } from 'src/interfaces/get-response.interface';
-import { ProductElement } from 'src/services/product.service';
+import { ProductElement } from 'src/interfaces/product-element.interface';
 import { getTotalPage } from './get-total-page.function';
 
 export function createProductPaginationForm(
